#121 Make tag validation depend on a log in
