+++ extension/ts/main/extensions/author/index.ts
@@ -1,3 +1,2 @@
 import "../../../../sass/author.sass";
 import "./authorPage";
-import "./bookPage";
